fix(schema): change duration field type from TINYINT to MEDIUMINT
- change duration field type from TINYINT to MEDIUMINT in the schema.sql file because the duration value can up to 60000 (60 seconds)
which exceeds the maximum value of TINYINT (255).

+++ docker/mysql/mysql-init-scripts/000_schema.sql
@@ -179,7 +179,7 @@ CREATE TABLE IF NOT EXISTS review_logs (
 
     learning_steps INT NOT NULL,
     state TINYINT NOT NULL,
-    duration TINYINT NOT NULL,
+    duration MEDIUMINT NOT NULL,
 
     PRIMARY KEY (user_id, id),
     INDEX ix_review_logs_usn (user_id ASC, usn ASC)
